UPGRADES
Become a Customer Rate Upgrade System Added

+++ Assets/EREN/ScriptsE/Customer Scripts/CustomerManager.cs
@@ -10,7 +10,7 @@ public class CustomerManager : MonoBehaviour
 
     public event Action OnHasAnyCustomer;
 
-    public int musteriOlmaSansii; //NPC'nin müþteri olma þansý. (Yüzdesel Olarak)
+    public float musteriOlmaSansii; //NPC'nin müþteri olma þansý. (Yüzdesel Olarak)
     public BakilacakYon bakilacakYon;
     public int sahnedeKacAdetParkPointVar;
     public int birYemekSansi;
